+++ app/src/main/java/to/bitkit/ui/components/Text.kt
@@ -24,6 +24,8 @@ fun Display(
     fontWeight: FontWeight = FontWeight.Black,
     fontSize: TextUnit = 44.sp,
     color: Color = MaterialTheme.colorScheme.primary,
+    maxLines: Int = Int.MAX_VALUE,
+    overflow: TextOverflow = if (maxLines == 1) TextOverflow.Ellipsis else TextOverflow.Clip,
     textAlign: TextAlign? = null,
 ) {
     Text(
@@ -33,8 +35,10 @@ fun Display(
             fontSize = fontSize,
             color = color,
         ),
+        maxLines = maxLines,
+        overflow = overflow,
         textAlign = textAlign,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -52,7 +56,7 @@ fun Display(
             lineHeight = fontSize,
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -67,7 +71,7 @@ fun Headline(
         style = AppTextStyles.Headline.merge(
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -85,7 +89,7 @@ fun Headline20(
             letterSpacing = (-.5).sp,
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -102,7 +106,7 @@ fun Headline24(
             lineHeight = 24.sp,
             color = color,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -123,7 +127,7 @@ fun Title(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -144,7 +148,7 @@ fun Subtitle(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -188,7 +192,7 @@ fun BodyM(
         maxLines = maxLines,
         minLines = minLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -228,7 +232,7 @@ fun BodyMSB(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -249,7 +253,7 @@ fun BodyMB(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -342,7 +346,7 @@ fun BodySB(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -359,7 +363,7 @@ fun Text13Up(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -380,7 +384,7 @@ fun Caption(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -437,7 +441,7 @@ fun Caption13Up(
             color = color,
             textAlign = textAlign,
         ),
-        modifier = modifier,
+        modifier = modifier
     )
 }
 
@@ -458,6 +462,6 @@ fun Footnote(
         ),
         maxLines = maxLines,
         overflow = overflow,
-        modifier = modifier,
+        modifier = modifier
     )
 }